fix: 修正动态声波 Swift 语法

+++ ios-app/LocalMusicPlayer/Features/NowPlaying/PlaybackWaveformView.swift
@@ -12,14 +12,14 @@ struct PlaybackWaveformView: View {
         ) { context in
             let time = context.date.timeIntervalSinceReferenceDate
             HStack(spacing: 4) {
-                ForEach(0..< twentyFourBars, id: \.self) { index in
+                ForEach(0..<twentyFourBars, id: \.self) { index in
                     let phase = time * 4.2 + Double(index) * 0.48
                     let amplitude = isPlaying
                         ? (0.25 + abs(sin(phase)) * 0.75)
                         : 0.24
                     Capsule()
                         .fill(PlayerTheme.accent.opacity(isPlaying ? 0.9 : 0.45))
-                        .frame(width: 3, height: 5 + amplitude * 20)
+                        .frame(width: 3, height: CGFloat(5 + amplitude * 20))
                 }
             }
             .frame(height: 28)
